Add source for jdk_1.4.2

+++ src/org/apache/xml/utils/res/XResources_cy.java
@@ -107,7 +107,7 @@ public class XResources_cy extends XResourceBundle
     { "numberGroups", new int[]{ 100, 10, 1 } },
 
     //These only used for mutiplicative-additive numbering
-    { "multiplier", new int[]{ 1000 } },
+    { "multiplier", new long[]{ 1000 } },
     { "multiplierChar", new char[]{ 0x03D9 } },
 
     // chinese only??
